feature benchmark: move Report and ReportMeasurement to separate file
